max software updated + threading tests

+++ pir_player_ext.py
@@ -1,4 +1,4 @@
-import subprocess
+import subprocess, datetime
 from gpiozero import DigitalInputDevice
 
 # Set the debug level
@@ -13,13 +13,15 @@ pir = DigitalInputDevice(26)
 #set the variables
 player_slow_playing = False 
 player_pir_playing = False
+player_slow_done = False
+player_pir_done = False
 
 player_slow_file = "/home/pi/Desktop/whos_afraid/player_slow.py"
 player_pir_file = "/home/pi/Desktop/whos_afraid/player_pir.py"
 
 while True:
     if not player_slow_playing and not player_pir_playing and not pir.value:
-        print("start player_slow")
+        print(f"{datetime.datetime.now().time()} start player_slow")
         player_slow_playing = True
         player_slow = subprocess.Popen(['python3', player_slow_file], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         player_slow_out, player_slow_err = player_slow.communicate()
@@ -29,18 +31,22 @@ while True:
             print('Error :', player_slow_err)        
             print('player_1', player_slow.pid)
             print('returncode', player_slow.returncode)
-    elif player_slow.returncode == 0:
-        print("killing pir player")
-        player_pir.kill()    
-        print("player_slow done")
+        if player_slow.returncode == 0:
+            player_slow_done = True
+    elif player_slow_playing and player_slow_done:
+        print(f"{datetime.datetime.now().time()}killing slow player")
+        player_slow.kill()    
+        print(f"{datetime.datetime.now().time()}player_slow done")
         player_slow_playing = False                
     elif pir.value and not player_pir_playing:
-        print("pir active")
+        print(f"{datetime.datetime.now().time()}pir active")
         player_pir_playing = True
-        print("killing slow player")
-        player_slow_playing = False
-        player_slow.kill()
-        print("start player_pir")
+        print(f"{datetime.datetime.now().time()}trying to kill slow player")
+        if player_slow_playing:
+            player_slow_playing = False
+            player_slow.kill()
+            print(f"{datetime.datetime.now().time()}killed slow player")
+        print(f"{datetime.datetime.now().time()}start player_pir")
         player_pir = subprocess.Popen(['python3', player_pir_file], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         player_pir_out, player_pir_err = player_pir.communicate()
         if DEBUG == 1:
@@ -48,9 +54,10 @@ while True:
             print('Output : ', player_pir_out)
             print('Error :', player_pir_err)
             print('returncode', player_pir.returncode)
-    elif player_pir.returncode == 0:
-        print("killing pir player")
+        if player_pir.returncode == 0:
+            player_pir_done = True            
+    elif player_pir_playing and player_pir_done:
+        print(f"{datetime.datetime.now().time()}killing pir player")
         player_pir.kill()
-        print("player_pir done")
+        print(f"{datetime.datetime.now().time()}player_pir done")
         player_pir_playing = False       
-
